chore: extract util to shared
